[Data] Move datasource/datasink tests into subfolder
Organize read/write tests under python/ray/data/tests/datasource and
update the HuggingFace helper import to match the new path.
