+++ python/ray/data/tests/datasource/test_binary.py
