+++ python/ray/data/tests/datasource/test_image.py
