+++ python/ray/data/tests/datasource/test_zip.py
